chore: fix lint warnings
