+++ src/context/AuthContext.ts
@@ -1,0 +1,22 @@
+import { User } from "@supabase/supabase-js";
+import { createContext, useContext } from "react";
+
+type AuthContextType = {
+  user: User | null;
+  loading: boolean;
+  signInWithMagicLink: (email: string) => Promise<void>;
+  signOut: () => Promise<void>;
+  getDisplayName: () => string;
+};
+
+export const AuthContext = createContext<AuthContextType | undefined>(
+  undefined,
+);
+
+export function useAuth() {
+  const context = useContext(AuthContext);
+  if (context === undefined) {
+    throw new Error("useAuth must be used within an AuthProvider");
+  }
+  return context;
+}
