+++ src/context/AuthProvider.tsx
@@ -1,16 +1,7 @@
 import { supabase } from "@/integrations/supabase/client";
+import { AuthContext } from "./AuthContext";
+import { useEffect, useState } from "react";
 import { User } from "@supabase/supabase-js";
-import { createContext, useContext, useEffect, useState } from "react";
-
-type AuthContextType = {
-  user: User | null;
-  loading: boolean;
-  signInWithMagicLink: (email: string) => Promise<void>;
-  signOut: () => Promise<void>;
-  getDisplayName: () => string;
-};
-
-const AuthContext = createContext<AuthContextType | undefined>(undefined);
 
 export function AuthProvider({ children }: { children: React.ReactNode }) {
   const [user, setUser] = useState<User | null>(null);
@@ -61,12 +52,4 @@ export function AuthProvider({ children }: { children: React.ReactNode }) {
       {children}
     </AuthContext.Provider>
   );
-}
-
-export function useAuth() {
-  const context = useContext(AuthContext);
-  if (context === undefined) {
-    throw new Error("useAuth must be used within an AuthProvider");
-  }
-  return context;
 }
