+++ src/context/BookingContext.ts
@@ -1,0 +1,39 @@
+import { Booking, User } from "@/types";
+import { createContext, useContext } from "react";
+
+interface BookingContextType {
+  bookings: Booking[];
+  user: User | null;
+  createBooking: (
+    booking: Omit<
+      Booking,
+      "id" | "createdAt" | "status" | "comments" | "createdBy"
+    > & { createdBy: User },
+  ) => Promise<string>;
+  addCommentToBooking: (
+    bookingId: string,
+    content: string,
+    email: string,
+  ) => Promise<string>;
+  getBookingById: (id: string) => Booking | undefined;
+  verifyBookingEmail: (id: string, token: string) => boolean;
+  verifyCommentEmail: (
+    bookingId: string,
+    commentId: string,
+    token: string,
+  ) => boolean;
+  approveBookingRequest: (id: string) => void;
+  getUserEmail: () => string | null;
+}
+
+export const BookingContext = createContext<BookingContextType | undefined>(
+  undefined,
+);
+
+export const useBooking = () => {
+  const context = useContext(BookingContext);
+  if (context === undefined) {
+    throw new Error("useBooking must be used within a BookingProvider");
+  }
+  return context;
+};
